Add LFH based system paths to search path stack
This might be a controversial change since it might be at the wrong location and maybe unwanted.
This pushes the standard system paths for the Linux to the bottom of the game file search paths.

This allows the files added with the build process to be added to a safe location with a package manager (`/usr/share/rbdoom3bfg`) and also makes easier to replace the files when building manually (`/usr/local/share/rbdoom3bfg`)